requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
pandas
dash
plotly
flask